feat: add pages on shell scripting
mainly bash

This was taken from Diane Fraser's 'Advanced Shell Scripting' guide. It
is heavily adapted to work well with markdown formatting. Some changes
were made for it to flow better and I took the liberty to split this up
into two pages (as things like regular expressions don't belong in the
basics).